collect all webdriver related method into base directory in core
