+++ lib/appium_lib/core/common.rb
@@ -1,9 +1,6 @@
 require_relative 'common/logger'
 require_relative 'common/error'
 require_relative 'common/log'
-require_relative 'common/wait'
-require_relative 'common/search_context'
 require_relative 'common/command'
 require_relative 'common/device'
 require_relative 'common/base'
-require_relative 'common/http_default'
