+++ lib/appium_lib/core/common/base/search_context.rb
